Refactor code to package classes
