+++ src/main/java/donna/Main.java
@@ -1,5 +1,6 @@
 package donna;
 
+import donna.gui.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
