+++ src/main/java/donna/gui/MainWindow.java
@@ -1,11 +1,12 @@
-package donna;
+package donna.gui;
 
+import donna.Donna;
+import donna.Ui;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 
 /**
